+++ .rsink/tools/pushover.sh
@@ -1,0 +1,16 @@
+#!/bin/sh
+# Send pushover notification
+
+appl="" # use your own
+user="" # use your own
+
+date=`date`
+
+curl -s \
+    -F "token=$appl" \
+    -F "user=$user" \
+    -F "title=rsink transfer complete" \
+    -F "message=$date" \
+    https://api.pushover.net/1/messages.json > /dev/null
+
+