+++ .rsink/constants.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-version="rsink 0.1+dev"
+version="rsink 0.2"
 config_file="config"
 profiles_directory="profiles"
 tools_directory="tools"